server/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.12.5
pydantic-settings==2.13.1
SQLAlchemy==2.0.36
asyncpg==0.30.0
python-dotenv==1.0.1
python-multipart==0.0.22
python-jose[cryptography]==3.5.0
bcrypt==5.0.0
fastapi-mail==1.6.2
typesense==2.0.0
email-validator==2.3.0
Authlib==1.6.9
httpx==0.28.1
